+++ src/Fields/field.jl
@@ -1,4 +1,25 @@
-using Oceananigans.Grids: show_domain
+using Base: @propagate_inbounds
+
+import Adapt
+using OffsetArrays
+
+import Oceananigans: datatuple
+using Oceananigans: @hascuda
+
+"""
+    AbstractField{A, G}
+
+Abstract supertype for fields stored on an architecture `A` and defined on a grid `G`.
+"""
+abstract type AbstractField{A, G} end
+
+"""
+    AbstractLocatedField{X, Y, Z, A, G}
+
+Abstract supertype for fields located at `(X, Y, Z)`, stored on an architecture `A`,
+and defined on a grid `G`.
+"""
+abstract type AbstractLocatedField{X, Y, Z, A, G} <: AbstractField{A, G} end
 
 """
     Cell
@@ -27,6 +48,9 @@ struct Field{X, Y, Z, A, G} <: AbstractLocatedField{X, Y, Z, A, G}
     end
 end
 
+Adapt.adapt_structure(to, field::Field{X, Y, Z}) where {X, Y, Z} =
+    Field{X, Y, Z}(adapt(to, data), field.grid)
+
 """
 	Field(L::Tuple, data::AbstractArray, grid)
 
@@ -36,21 +60,20 @@ of length 3 whose elements are `Cell` or `Face`.
 Field(L::Tuple, data::AbstractArray, grid) = Field{L[1], L[2], L[3]}(data, grid)
 
 """
-    Field(L::Tuple, arch::AbstractArchitecture, grid)
+    Field(L::Tuple, arch, grid)
 
 Construct a `Field` on architecture `arch` and `grid` at location `L`,
 where `L` is a tuple of `Cell` or `Face` types.
 """
-Field(L::Tuple, arch::AbstractArchitecture, grid) =
-    Field{L[1], L[2], L[3]}(zeros(arch, grid), grid)
+Field(L::Tuple, arch, grid) = Field{L[1], L[2], L[3]}(zeros(arch, grid), grid)
 
 """
-    Field(X, Y, Z, arch::AbstractArchitecture, grid)
+    Field(X, Y, Z, arch, grid)
 
 Construct a `Field` on architecture `arch` and `grid` at location `X`, `Y`, `Z`,
 where each of `X, Y, Z` is `Cell` or `Face`.
 """
-Field(X, Y, Z, arch::AbstractArchitecture, grid) =  Field((X, Y, Z), arch, grid)
+Field(X, Y, Z, arch, grid) =  Field((X, Y, Z), arch, grid)
 
 """
     CellField([T=eltype(grid)], arch, grid)
@@ -100,6 +123,7 @@ architecture(o::OffsetArray) = architecture(o.parent)
 
 @propagate_inbounds getindex(f::Field, inds...) = @inbounds getindex(f.data, inds...)
 @propagate_inbounds setindex!(f::Field, v, inds...) = @inbounds setindex!(f.data, v, inds...)
+
 @inline lastindex(f::Field) = lastindex(f.data)
 @inline lastindex(f::Field, dim) = lastindex(f.data, dim)
 
@@ -152,131 +176,3 @@ const AbstractCPUField =
 
 @hascuda const AbstractGPUField =
     AbstractField{A, G} where {A<:OffsetArray{T, D, <:CuArray} where {T, D}, G}
-
-set!(u::Field, v::Number) = @. u.data = v
-
-set!(u::Field{X, Y, Z, A}, v::Field{X, Y, Z, A}) where {X, Y, Z, A} =
-    @. u.data.parent = v.data.parent
-
-"Set the CPU field `u` to the array `v`."
-function set!(u::AbstractCPUField, v::Array)
-    for k in 1:u.grid.Nz, j in 1:u.grid.Ny, i in 1:u.grid.Nx
-        u[i, j, k] = v[i, j, k]
-    end
-    return nothing
-end
-
-# Set the GPU field `u` to the array `v`.
-@hascuda function set!(u::AbstractGPUField, v::Array)
-    v_field = Field(location(u), CPU(), u.grid)
-    set!(v_field, v)
-    set!(u, v_field)
-    return nothing
-end
-
-# Set the GPU field `u` to the CuArray `v`.
-@hascuda function set!(u::AbstractGPUField, v::CuArray)
-    @launch device(GPU()) config=launch_config(u.grid, :xyz) _set_gpu!(u.data, v, u.grid)
-    return nothing
-end
-
-function _set_gpu!(u, v, grid)
-    @loop for k in (1:grid.Nz; (blockIdx().z - 1) * blockDim().z + threadIdx().z)
-        @loop for j in (1:grid.Ny; (blockIdx().y - 1) * blockDim().y + threadIdx().y)
-            @loop for i in (1:grid.Nx; (blockIdx().x - 1) * blockDim().x + threadIdx().x)
-	            u[i, j, k] = v[i, j, k]
-	        end
-	    end
-    end
-    return nothing
-end
-
-# Set the GPU field `u` data to the CPU field data of `v`.
-@hascuda set!(u::AbstractGPUField, v::AbstractCPUField) = copyto!(u.data.parent, v.data.parent)
-
-# Set the CPU field `u` data to the GPU field data of `v`.
-@hascuda set!(u::AbstractCPUField, v::AbstractGPUField) = u.data.parent .= Array(v.data.parent)
-
-"Set the CPU field `u` data to the function `f(x, y, z)`."
-set!(u::AbstractCPUField, f::Function) = interior(u) .= f.(nodes(u)...)
-
-# Set the GPU field `u` data to the function `f(x, y, z)`.
-@hascuda function set!(u::AbstractGPUField, f::Function)
-    u_cpu = Field(location(u), CPU(), u.grid)
-    set!(u_cpu, f)
-    set!(u, u_cpu)
-    return nothing
-end
-
-"""
-    set!(model; kwargs...)
-
-Set velocity and tracer fields of `model`. The keyword arguments
-`kwargs...` take the form `name=data`, where `name` refers to one of the
-fields of `model.velocities` or `model.tracers`, and the `data` may be an array,
-a function with arguments `(x, y, z)`, or any data type for which a
-`set!(ϕ::AbstractField, data)` function exists.
-
-Example
-=======
-```julia
-model = Model(grid=RegularCartesianGrid(size=(32, 32, 32), length=(1, 1, 1))
-
-# Set u to a parabolic function of z, v to random numbers damped
-# at top and bottom, and T to some silly array of half zeros,
-# half random numbers.
-
-u₀(x, y, z) = z/model.grid.Lz * (1 + z/model.grid.Lz)
-v₀(x, y, z) = 1e-3 * rand() * u₀(x, y, z)
-
-T₀ = rand(size(model.grid)...)
-T₀[T₀ .< 0.5] .= 0
-
-set!(model, u=u₀, v=v₀, T=T₀)
-```
-"""
-function set!(model::AbstractModel; kwargs...)
-    for (fldname, value) in kwargs
-        if fldname ∈ propertynames(model.velocities)
-            ϕ = getproperty(model.velocities, fldname)
-        elseif fldname ∈ propertynames(model.tracers)
-            ϕ = getproperty(model.tracers, fldname)
-        else
-            throw(ArgumentError("name $fldname not found in model.velocities or model.tracers."))
-        end
-        set!(ϕ, value)
-    end
-    return nothing
-end
-
-function set!(Φ::NamedTuple; kwargs...)
-    for (fldname, value) in kwargs
-        ϕ = getproperty(Φ, fldname)
-        set!(ϕ, value)
-    end
-    return nothing
-end
-
-set_ic!(model; kwargs...) = set!(model; kwargs...) # legacy wrapper
-
-Adapt.adapt_structure(to, field::Field{X, Y, Z}) where {X, Y, Z} =
-    Field{X, Y, Z}(adapt(to, data), field.grid)
-
-show_location(X, Y, Z) = string("(", string(typeof(X())), ", ",
-                                     string(typeof(Y())), ", ",
-                                     string(typeof(Z())), ")")
-
-show_location(field::AbstractLocatedField{X, Y, Z}) where {X, Y, Z} = show_location(X, Y, Z)
-
-short_show(a) = string(typeof(a))
-shortname(a::Array) = string(typeof(a).name.wrapper)
-
-show(io::IO, field::Field) =
-    print(io,
-          short_show(field), '\n',
-          "├── data: ", typeof(field.data), '\n',
-          "└── grid: ", typeof(field.grid), '\n',
-          "    ├── size: ", size(field.grid), '\n',
-          "    └── domain: ", show_domain(field.grid), '\n')
-
-short_show(field::AbstractLocatedField) = string("Field at ", show_location(field))
